+++ packages/core/src/index.ts
@@ -1,3 +1,6 @@
+export type {
+  FileMetadata, UploadHeaders,
+} from './functions/upload.js'
 export * from './accessControl.js'
 export * from './assets.js'
 export * from './collection/index.js'